+++ packages/remix-dev/__tests__/create-test.ts
@@ -8,7 +8,12 @@ import stripAnsi from "strip-ansi";
 
 import { run } from "../cli/run";
 import { server } from "./msw";
-import { flatRoutesWarning } from "../config";
+import {
+  errorBoundaryWarning,
+  flatRoutesWarning,
+  formMethodWarning,
+  metaWarning,
+} from "../config";
 
 beforeAll(() => server.listen({ onUnhandledRequest: "error" }));
 afterAll(() => server.close());
@@ -348,7 +353,13 @@ describe("the create command", () => {
       "--no-typescript",
     ]);
     expect(output.trim()).toBe(
-      flatRoutesWarning +
+      errorBoundaryWarning +
+        "\n" +
+        formMethodWarning +
+        "\n" +
+        metaWarning +
+        "\n" +
+        flatRoutesWarning +
         "\n\n" +
         getOptOutOfInstallMessage() +
         "\n\n" +